fix types

+++ blaze/evaluator/simulator/request_queue.py
@@ -30,7 +30,7 @@ class Node(NamedTuple):
 class QueueItem:
     """ An item in the RequestQueue """
 
-    def __init__(self, node: Node, size: int, origin: str, delay_ms: int = 0):
+    def __init__(self, node: Node, size: int, origin: str, delay_ms: float = 0):
         self.node = node
         self.bytes_left = size
         self.origin = origin
@@ -97,7 +97,7 @@ class RequestQueue:
         self.queue = [qi for qi in self.queue if qi.node != node]
         self.delayed = [qi for qi in self.delayed if qi.node != node]
 
-    def add_with_delay(self, node: Node, delay_ms: int):
+    def add_with_delay(self, node: Node, delay_ms: float):
         """
         Adds an item to the queue but does not start it until the delay has occurred. Additionally,
         this method checks to see if a connection has been opened for the resource's origin. If not,
@@ -109,7 +109,7 @@ class RequestQueue:
         if domain not in self.connected_origins:
             num_rtts += 1
 
-        delay_ms = max(0, delay_ms + (num_rtts * self.rtt_latency_ms))
+        delay_ms = max(0.0, delay_ms + (num_rtts * self.rtt_latency_ms))
         queue_item = QueueItem(node, node.resource.size, domain, delay_ms)
         if delay_ms <= 0:
             self.queue.append(queue_item)
